+++ CHANGELOG.md
@@ -3,8 +3,9 @@ Who knows what the future holds...
 
 # 0.X.Y - DD/MM/YYYY
 ### Changes:
-- Added [Valheim](https://store.steampowered.com/app/892970/Valheim/) support.
-- Added [The Front](https://store.steampowered.com/app/2285150/The_Front/) support.
+- [Valheim](https://store.steampowered.com/app/892970/Valheim/) support.
+- [The Front](https://store.steampowered.com/app/2285150/The_Front/) support.
+- [Conan Exiles](https://store.steampowered.com/app/440900/Conan_Exiles/) support.
 
 ### Breaking:
 Game: